% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_CFS_light.R
\name{CFS_light_save}
\alias{CFS_light_save}
\title{CFS light save}
\usage{
CFS_light_save(data = NULL, assay = "SCT")
}
\arguments{
\item{data}{Seurat object}

\item{assay}{assay to keep for visualization}
}
\value{
Light Seurat object for visualization with the shiny tool
}
\description{
Create a Seurat object in a lighter format for visualization with the Shiny tool
}
\examples{
light_data <- CFS_light_save(data=NULL, assay = 'SCT')

}
