% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cluster_ICA.R
\name{Cluster_ICA}
\alias{Cluster_ICA}
\title{Cluster ICA}
\usage{
Cluster_ICA(data = NULL, ICs = c(1), res = 1.2, spread = 3)
}
\arguments{
\item{data}{Seurat object to clusterise}

\item{ICs}{Vector of IC numbers to use for the clusterisation and UMAP}

\item{res}{Resolution of the clusterisation}
}
\value{
Seurat object with clustered cells and UMAP
}
\description{
Clusters cells and run UMAP
}
\examples{
data <- Cluster_ICA(data=data,ICs=c(1,2,3,4,5,6,7,8,9),res=1.2)
}
