% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_CosMX.R
\name{Create_CosMX_seurat}
\alias{Create_CosMX_seurat}
\title{Create_vizgen_seurat}
\usage{
Create_CosMX_seurat(tx = NULL, pixel_format = 600, min.features = 5)
}
\arguments{
\item{tx}{table of transcripts}

\item{pixel_format}{size of the binning (in px)}

\item{min.features}{minimum number of transcripts to keep per spot.}
}
\value{
Seurat object corresponding to the binned CosMX data
}
\description{
Convert vizgen data using grid based method
}
\examples{
data <- Create_CosMX_seurat(tx=spot_table, pixel_format = 40, min.features = 5)

}
