% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_Vizgen.R
\name{Create_vizgen_seurat}
\alias{Create_vizgen_seurat}
\title{Create_vizgen_seurat}
\usage{
Create_vizgen_seurat(
  spot_gene_expression = NULL,
  pixel_format = 40,
  min.features = 5
)
}
\arguments{
\item{spot_gene_expression}{table of transcripts}

\item{pixel_format}{size of the binning (in px)}

\item{min.features}{minimum number of transcripts to keep per spot.}
}
\value{
Seurat object corresponding to the binned vizgen data
}
\description{
Convert vizgen data using grid based method
}
\examples{
data <- Create_vizgen_seurat(spot_gene_expression=spot_table, pixel_format = 40, min.features = 5)

}
