% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoHeatmapICA.R
\name{DoHeatmapICA}
\alias{DoHeatmapICA}
\title{Do Heatmap for the complete ICA analysis}
\usage{
DoHeatmapICA(data = "NULL", nics = "IC_1", ngenes = 10)
}
\arguments{
\item{data}{Seurat object to analyse}

\item{nics}{ICs to add to the heatmap}

\item{ngenes}{Number of genes per IC to add to the heatmap}
}
\value{
A data seurat object containing the heatmap necessary full heatmap for the shiny tool
}
\description{
Prepare full ICs heatmap for Shiny analysis
}
\examples{
data <- DoHeatmapICA(data=data,nics=names(which(data\@misc$GeneAndStat$Kurtosis_ICs>3)),ngenes=10)
}
