% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunICA.R
\name{ICASpatial}
\alias{ICASpatial}
\title{ICA Spatial}
\usage{
ICASpatial(
  data = NULL,
  ncis = 100,
  maxit = 600,
  method = "icafast",
  sd = 3,
  ...
)
}
\arguments{
\item{data}{Seurat object to analyse}

\item{ncis}{Number of independant component to separate in the sample}

\item{maxit}{Number of iterations}

\item{method}{Method to use between "icafast", "icaimax", "icajade"}

\item{sd}{Value of the kurtosis filter on IC}
}
\value{
An analysed Seurat object
}
\description{
Complete ICA analysis on the spatial sample after normalisation and regression. Then correct sign and determine stats from IC and Genes
}
\examples{
data = ICASpatial(data=data,ncis=100,maxit=600,method="icafast",sd=3)

}
