% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICGeneAndStats.R
\name{ICGeneAndStats}
\alias{ICGeneAndStats}
\title{ICs Stats}
\usage{
ICGeneAndStats(data = "NULL", sd = 3)
}
\arguments{
\item{data}{Seurat object to analyse}

\item{sd}{standard deviation contribution to take into account}
}
\value{
Adds a list of analysis of ICs to the object
}
\description{
Adds Statistics analysis of ICs to the object
}
\examples{
data <- ICGeneAndStats(data=data,sd=3)
}
