% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchCFS.R
\name{launchCFS}
\alias{launchCFS}
\title{Launch Shiny}
\usage{
launchCFS(
  mode = "open",
  maxFileSize = 15000,
  file_to_load = NULL,
  welcome_message = NULL,
  projections_show_hover_info = TRUE,
  ...
)
}
\value{
None
}
\description{
Start the shiny for visualization, analysis and annotation purposes
}
\examples{
launchCFS()

}
