% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchShiny.R
\name{launchShiny}
\alias{launchShiny}
\title{Launch Shiny}
\usage{
launchShiny(
  mode = "open",
  maxFileSize = 15000,
  file_to_load = NULL,
  welcome_message = NULL,
  projections_show_hover_info = TRUE,
  ...
)
}
\value{
None
}
\description{
Start the shiny for visualization and annotation purposes
}
\examples{
launchShiny()

}
