% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copykat.R
\name{Ploidie_search}
\alias{Ploidie_search}
\title{Ploidie search for cancer cells. (Non functionnal)}
\usage{
Ploidie_search(
  data = NULL,
  copykat.test = NULL,
  species = "S",
  threads = 4,
  kcut = 2,
  annotate = TRUE,
  genome = "hg20"
)
}
\arguments{
\item{data}{Seurat object to analyse}

\item{species}{???}

\item{threads}{Number of threads to use}

\item{kcut}{???}

\item{annotate}{chose whether or not to annotate the analysis, of return the Copykat object.}

\item{genome}{Genome to use for the analysis, either "hg20" or "mm"}
}
\value{
A list of analysis object
}
\description{
Search for ploidie alteration within the sample.
}
\examples{
data <- Ploidie_search(data=data, species="S", threads=4, kcut=2, annotate=TRUE, genome="hg20")
}
