% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{PrepNormData}
\alias{PrepNormData}
\title{Normalize data}
\usage{
PrepNormData(
  data = NULL,
  organism = "Hs",
  variable_features = 20000,
  min_cells = 5
)
}
\arguments{
\item{data}{Seurat object to analyse}

\item{organism}{"Hs" (human) or "Mm" (mice)}

\item{variable_features}{Number of variable features to use for the SCTransform}

\item{min_cells}{Filters genes less present than the indicated number of spots to use for the SCTransform}
}
\value{
an object with normalized data
}
\description{
Determine mitochondrial ratio, hb ratio and normalise using SCTransform
}
\examples{
data <- PrepNormData(data=data,organism="Hs",variable_features=20000, min_cells = 5)
}
