% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_IC_and_Enrich.R
\name{Show_IC_and_Enrich}
\alias{Show_IC_and_Enrich}
\title{enrich IC}
\usage{
Show_IC_and_Enrich(
  data = NULL,
  dbs = c("GO_Biological_Process_2015"),
  kurtosis = 3
)
}
\value{
datatables of enrichment intro the seurat object
}
\description{
Search and enrich IC genes using EnrichR. Result saved in misc$IC_X$en
}
\examples{
data = Show_IC_and_Enrich(data=data,dbs=c("GO_Biological_Process_2015"))

}
