% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_Cluster_ICA.R
\name{sub_Cluster_ICA}
\alias{sub_Cluster_ICA}
\title{Fahrenheit conversion}
\usage{
sub_Cluster_ICA(
  data = NULL,
  ICs = c(1),
  ncis = 30,
  organism = "Hs",
  res = 1.2,
  variable_features = 20000
)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- F_to_C(50);
temp2 <- F_to_C( c(50, 63, 23) );
}
