% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich_ICA.R
\name{Enrich_ICA}
\alias{Enrich_ICA}
\title{enrich IC}
\usage{
Enrich_ICA(
  data = NULL,
  dbs = c("PanglaoDB_Augmented_2021", "CellMarker_2024", "Tabula_Sapiens",
    "Tabula_Muris", "GO_Biological_Process_2023", "GO_Molecular_Function_2023",
    "GO_Cellular_Component_2023", "MSigDB_Hallmark_2020", "Reactome_2022",
    "KEGG_2021_Human"),
  kurtosis = 3,
  overwrite = TRUE,
  enrich_positive_genes = TRUE
)
}
\arguments{
\item{data}{Seurat object to clusterise}

\item{dbs}{names of enrichR compatible databases to use}

\item{kurtosis}{kurtosis filter value to use}

\item{overwrite}{overwrite preexisting annotations in the object.}
}
\value{
datatables of enrichment into the seurat object
}
\description{
Search and enrich IC genes using EnrichR. Result saved in misc$IC_X$en
}
\examples{
data = Enrich_ICA(data=data,dbs=c("GO_Biological_Process_2015"), kurtosis = 3, overwrite = TRUE)

}
