% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrich_ICA.R
\name{Enrich_ICA}
\alias{Enrich_ICA}
\title{enrich IC}
\usage{
Enrich_ICA(
  data = NULL,
  dbs = c("GO_Biological_Process_2015"),
  kurtosis = 3,
  overwrite = TRUE
)
}
\value{
datatables of enrichment intro the seurat object
}
\description{
Search and enrich IC genes using EnrichR. Result saved in misc$IC_X$en
}
\examples{
data = Enrich_ICA(data=data,dbs=c("GO_Biological_Process_2015"), kurtosis = 3, overwrite = TRUE)

}
